Add files via upload
To assess Turkmenistan's progress in achieving the Sustainable Development Goals, the outcomes of SDGs 1–8 in 2019 were shown in a horizontal bar chart (SDGs). A horizontal bar chart was a great choice as it's easy to compare the outcomes. The measure name appears in the row, and the measured value appears in the column, for each SDG measure value that was aggregated. To restrict the results to 2019 and Turkmenistan, the variables for year and country, and measure names were selected in the filter section. The resulting graphic revealed that SDG 7 had the highest score reached in 2019 while SDG 4 had the lowest score. The method used made it possible to see the progress Turkmenistan achieved toward reaching the SDGs in 2019, allowing academics to pinpoint the country's development projects' strong and weak points.